% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.mixEM}
\alias{logLik.mixEM}
\title{Log-Likelihood of \code{'mixEM'} Object}
\usage{
\method{logLik}{mixEM}(object, ...)
}
\arguments{
\item{object}{\code{'mixEM'} object, currently only the returned value of
\link[mixtools]{normalmixEM} and \link[mixtools]{gammamixEM} are supported}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=logLik.mixEM]{logLik.mixEM()}} returns a \link[stats]{logLik} object.
}
\description{
To obtain the log-Likelihood of \code{'mixEM'} object, based on \CRANpkg{mixtools} 2020-02-05.
}
\keyword{internal}
