% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort.mixEM}
\alias{sort.mixEM}
\title{Sort \code{'mixEM'} Object by First Parameters}
\usage{
\method{sort}{mixEM}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{\code{'mixEM'} object}

\item{decreasing}{\link[base]{logical} scalar. Should the sort by \eqn{mu}'s
be increasing (\code{FALSE}, default) or decreasing (\code{TRUE})?}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=sort.mixEM]{sort.mixEM()}} returns a \code{'mixEM'} object.
}
\description{
To \link[base]{sort} a \code{'mixEM'} object by its first parameters, i.e.,
\eqn{\mu}'s for normal mixture, \eqn{\alpha}'s for \eqn{\gamma}-mixture, etc.
}
\details{
\link[mixtools]{normalmixEM} does \emph{not} order the location parameter
}
\keyword{internal}
